feat: 棘轮式质量门禁 + P1 Bug 修复 + 统一打包
## 棘轮质量门 (Ratchet Gate)

gsimap / gsim-app: Compiler -Werror + Spotless + Tests + SpotBugs(0 bugs) = 铁门
gsim-lib: 同上 + SpotBugs 基线化(398条) + check-baseline.sh 禁止新增

## P1 Bug 修复

- MapService.evict(): Canvas 缓存键 worldId → worldId:nodeId 前缀匹配
- MapService.persistBlocks(): 移除 saveMap() 之后的重复 evict()
- MapResolver.history(): 添加 chain.isEmpty() 空链守卫

## 统一打包

- 移除 gsimap maven-shade-plugin → thin jar
- gsim-app 为唯一 fat jar 产出
- 删除 dependency-reduced-pom.xml + 加入 .gitignore

## CI / 编译器

- CI: mvn test+package → mvn verify, 触发 branches: ["**"]
- 编译器: -Xlint:all,-processing,-options -Werror
- Checkstyle: 移除与 Spotless 重叠的 15 条格式规则

## SpotBugs 零容忍 (gsimap)

- DMI_RANDOM_USED_ONLY_ONCE ×2 → ThreadLocalRandom
- MS_PKGPROTECT → HEX_DIRS package-private
- UC_USELESS_OBJECT → 移除无用 tags 变量

## gsim-lib 存量警告修复

- AppConfig: @SuppressWarnings("this-escape")
- LlmConfigManager: 补充 @Deprecated
- ApiRouter: 消除 deprecated API + redundant cast
- ImportDocumentService/UnknownToolException: serialVersionUID
- CliWebSocketServer/CliWsProgressSink: redundant cast

## 新 Skill

.claude/skills/gsim-integration-test/ — 集成验收测试(质量门+MCP+CRUD+权限+边界)

+++ gsim-lib/src/main/java/com/gsim/webui/CliWebSocketServer.java
@@ -482,7 +482,7 @@ public class CliWebSocketServer {
             out.write(len & 0xff);
         } else {
             out.write(127);
-            for (int i = 7; i >= 0; i--) out.write((int) ((len >> (i * 8)) & 0xff));
+            for (int i = 7; i >= 0; i--) out.write(((len >> (i * 8)) & 0xff));
         }
 
         out.write(payload);
